Refactor: Remove Winget and File services, and related models
- Deleted WingetOperationOptions and WingetOperationResult models as they are no longer needed.
- Removed WingetService which handled backup and restore operations for Winget packages.
- Deleted FileService which provided file operations including combining files.
- Updated DatabaseService to remove backup and restore methods, focusing on database listing and connection testing.
- Introduced DatabaseConnectionOptions model to encapsulate database connection parameters.
- Updated tests to remove dependencies on removed services.
