--- FurLab.Core/Models/FileOperationOptions.cs
@@ -1,32 +1,0 @@
-namespace FurLab.Core.Models;
-
-/// <summary>
-/// Represents options for a file combine operation.
-/// </summary>
-public record FileCombineOptions
-{
-    /// <summary>
-    /// Gets or sets the input file or directory path.
-    /// </summary>
-    public string Input { get; init; } = string.Empty;
-
-    /// <summary>
-    /// Gets or sets the output file path.
-    /// </summary>
-    public string? Output { get; init; }
-
-    /// <summary>
-    /// Gets or sets the file pattern to match (e.g., "*.txt").
-    /// </summary>
-    public string? Pattern { get; init; }
-
-    /// <summary>
-    /// Gets or sets whether to include subdirectories.
-    /// </summary>
-    public bool Recursive { get; init; }
-
-    /// <summary>
-    /// Gets or sets the separator to use between files.
-    /// </summary>
-    public string? Separator { get; init; }
-}
